adding unit tests

+++ cvd_image_converter/__main__.py
@@ -2,12 +2,14 @@
 
 
 import argparse
+import os
 import cv2
 from .cvd_image_converter import convert_color_palette
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
+        prog='python -m cvd_image_converter',
         description='Convert an image to a CVD safe color space')
     parser.add_argument(
         'input_image',
@@ -18,5 +20,7 @@ if __name__ == '__main__':
         action='store',
         help='Output image, CVD safe')
     args = parser.parse_args()
+    if not os.path.isfile(args.input_image):
+        raise Exception("Input image doesn't exist, please check the path.")
     input_image = cv2.imread(args.input_image)
     cv2.imwrite(args.output_image, convert_color_palette(input_image))
